BUG: Make sure numpy globals keep identity after reload.
Reloading currently causes problems because global classes defined in
numpy/__init__.py change their identity (a is b) after reload. The
solution taken here is to move those classes to a new non-reloadable
module numpy/_globals and import them into numpy from there.

Classes moved are ModuleDeprecationWarning, VisibleDeprecationWarning,
and _NoValue.

Closes #7844.
